cleaned up boundaries between matrix and graph operations
